requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Audio Transcription
transformers>=4.40.0
accelerate>=0.27.0
numpy

# Audio Capture & Speech Detection
pyaudio
webrtcvad
SpeechRecognition

# HTTP Client
requests

# Speaker Diarization
torch>=2.0.0
torchvision
torchaudio
pyannote.audio
noisereduce

# Audio Processing
pydub

# Production Server
gunicorn==21.2.0
